audio: Move tag helpers to mopidy.audio.tags
